Huge reformat because the previous architecture did not make use of the WindowApp model and thus it was not as good
Though, I feel like I'm doing something very bad at having the Window basic info like name and icon, in the same place as the composable lambdas

Even though, I'm tired. I just want to sleep and I will leave it like that. We will see if at the end it causes a major performance issue, and if so, then would be time to fix it.