Finished settings app at 100%! Yey!
I'm are almost half way in. Remaining apps are the web browser, audio player, photo viewer and the file explorer.

I think the most complex one will be the file explorer (due to the navigation features, I'm not sure if I will hardcode paths), and the second one will (probably) be the web browser due how feature-rich is the UI of edge

Anyways! Testing if I can do a newline in here using \n Is this in a new line?