Apps renamed, window detection system updated & started implementation of Music Player App
Opened windows (or all windows related in general) was handled by strings, which was a pretty bad decision but that worked on early stages. Now that I renamed the apps, a big mess was caused by these strings that now had to be replace everywhere.

That's why, instead, I decided to create an enum class called AppId