Several small improvements in Media Player app
Moved playerState: MediaPlayerState instantiation from MediaPlayerWindowContent to MediaPlayerApp, so when the app is 'closed' then the audio can be stopped, simulating the real behaviour of a media player.

Merged the code of changeAudio into a single function, which it was playToggler, but it was renamed to controlMedia, which is more descriptive.

I also added a new function: PlayerControls.STOP to clear all when the app exists.

The previously mentioned bug of having audio play not synced with the GUI is still here. I'll later tackle down that.