Fixed media player app desktop target synchronization issue
I was doing a play() after setSource(), but seems like it was executing way to fast, and setSource(), as being executed in another thread, was not setting the media ready by the moment of play(). So play() was playing... nothing, when called. Time to use the isBuffering and add a visual element that helps to make the user know at web target that the content is loading (maybe a circular progress bar)