Finished implementing the media player bar
It is not polished as I would like it to be. For example, the drag n drop feature in the thumb is a little bit buggy, and the trying to play-stop-play again in Desktop is literally broken

meh, anyways. The media player is almost completed. I just need to find a way to fix the audios from WASM (it reproduces audio from online, but not from local)