Small changes on web browser and fixed bug in media player
I finally understood the issue of the progress bar thumb feeling a bit jammed when drag and dropping. In summary, it was a conflict between user interaction and the UI (between on drag event and the LaunchedEffect code).

It was easily fixed after adding a flag isDragging, and controling it from the detectDragGestures

But now that I learned about side effects and the effect APIs... I wonder how horrid is the rest of the code I have written so far...