Early implementation of Homepage on Web browser
Dividers where extracted to its own composable, and a vertical scrollbar to scroll was created. This bar should work with lazycolumns and vertical grids (that I'll use in the gallery section).

Posts are ready to be created... or well, I need to create a header and a footer composable first.